Moved login stuff to new module login.py
